Move get_list_or_len to from chia.plot_sync.receiver
